+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 1.0.17
 
+- Add option `set_order` in `get_sig_similarity()` (#274).
 - Outputed more stats information in `output_sig()`.
 - Fixed default y axis title in `show_sig_bootstrap_error()`, now it is "Reconstruction error (L2 norm)"